pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python3-nmap>=1.9.1",
    "pdfkit>=1.0.0",
    "beautifulsoup4>=4.13.4",
    "flask-login>=0.6.3",
    "pyopenssl>=25.0.0",
    "cryptography>=44.0.3",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "whois>=1.20240129.2",
    "reportlab>=4.4.0",
    "werkzeug>=3.1.3",
    "python-whois>=0.9.5",
    "trafilatura>=2.0.0",
    "pymysql>=1.1.1",
]
